+++ tests/test_osmnx.py
@@ -26,6 +26,7 @@ from shapely.geometry import MultiPoint
 from shapely.geometry import MultiPolygon
 from shapely.geometry import Point
 from shapely.geometry import Polygon
+from xml.etree import ElementTree as etree
 
 import osmnx as ox
 
@@ -182,6 +183,25 @@ def test_osm_xml():
     df = pd.DataFrame({"u": [54, 2, 5, 3, 10, 19, 20], "v": [76, 3, 8, 10, 5, 20, 15]})
     ordered_nodes = ox.osm_xml._get_unique_nodes_ordered_from_way(df)
     assert ordered_nodes == [2, 3, 10, 5, 8]
+
+    # test roundabout handling
+    overpass_date_str = '[date:"2023-04-01T00:00:00Z"]'
+    ox.settings.overpass_settings += overpass_date_str
+    ox.settings.bidirectional_network_types = []
+    g = ox.graph_from_point(
+        (39.09091886432667, -84.52266036019037),
+        dist=int(np.floor(1609.34 * (5.5))),
+        dist_type="bbox",
+        network_type="drive",
+        simplify=False,
+        retain_all=False,
+    )
+    nodes, edges = ox.graph_to_gdfs(g)
+    way_df = edges[edges["osmid"] == 570883705]  # roundabout
+    root = etree.Element("osm", attrib={"version": "0.6", "generator": "OSMnx"})
+    first = way_df.iloc[0].dropna().astype(str)
+    edge = etree.SubElement(root, "way")
+    ox.osm_xml._append_nodes_as_edge_attrs(edge, first, way_df)
 
     ox.settings.all_oneway = default_all_oneway